+++ index.php
@@ -7,8 +7,6 @@ use Dotenv\Dotenv;
 $dotenv = Dotenv::createImmutable(__DIR__);
 $dotenv->load();
 
-
-
 use App\Controllers\NewsController;
 use App\Controllers\CommentController;
 use App\Repositories\NewsRepository;
@@ -17,7 +15,6 @@ use App\Services\Connections\DatabaseConnectionService;
 use App\Services\DatabaseService;
 
 $database_config = require_once "./config/database.php";
-
 
 $databaseConnection = new DatabaseConnectionService($database_config['mysql']);
 $databaseService = new DatabaseService($databaseConnection);
@@ -30,7 +27,7 @@ foreach (NewsController::getInstance($newsRepository, $commentRepository)->listN
 	echo ($news->getBody() . "\n");
 	foreach (CommentController::getInstance($commentRepository)->listComments() as $comment) {
 		if ($comment->getNewsId() == $news->getId()) {
-			echo ("Comment " . $comment->getId() . " : " . $comment->getBody() . "\n");
+			echo ("Comment " . $comment->getId() . " : " . $comment->getBody() . "\/n");
 		}
 	}
 }